pyproject.toml
[project]
name = "api-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
